refactor(open-mpm): split skills/ oversized files (closes #363)
Split 4 oversized files in skills/ under the 500-line cap. Pure
reorganization — no logic changes, public API preserved via re-exports:

- mod.rs (1511) -> mod.rs (36) + types.rs + loader.rs + llm.rs + mod_tests.rs
- registry.rs (1432) -> registry/{mod,meta,scan,tests,tests_persistence}.rs
- sources.rs (643) -> sources/{mod,tests}.rs
- global_cache.rs (568) -> global_cache/{mod,tests}.rs

All files now under 500 lines. The flat skills::* public surface
(SkillRegistry, SkillsLoader, SkillEntry, strip_frontmatter,
select_skills_via_llm, skill_llm_enabled) is preserved through re-exports.

Part of #356
